Update CsvDataWriterOptions.cs (#233)
The newline default is "\n", not Environment.NewLine()
This is probably as it should be, as it would be awkward for results to change based on what system runs the software (e.g. in mixed Windows/Unix server environments), and because all software that matters, including Excel, support \n by default.